Fix scan float32, flag lengths, dict update mode, and solver accounting
scan_cycle_reduction rejected float32 at graph build, make_not_loglin_flags returned the wrong length with log_linearize=False, check_bk_condition returned numpy.bool_, SymbolDictionary.update(**kwargs) bypassed the mode check, and the sparse root solvers validated a list x0 before casting and undercounted evaluations on a fatal rejection.

+++ gEconpy/solvers/cycle_reduction.py
@@ -444,7 +444,7 @@ def _scan_cycle_reduction(A, B, C, max_iter: int = 1000, tol: float = 1e-7) -> l
     n = A.shape[0]
     idx_0 = pt.arange(n)
     idx_1 = idx_0 + n
-    norm = np.array(1e9, dtype="float64")
+    norm = np.array(1e9, dtype=A.dtype)
     step_num = pt.zeros((), dtype="int32")
     *_, A1_hat, norm, n_steps = pytensor.scan(
         step,
